avbtool: add_hashtree_footer: Add --calc_max_image_size option.
In the Android build system system.img is built in a way so it takes up
the full partition leaving no free space at the end (unlike ChromeOS
where it's resized). We therefore need a way to know how much space
we're going to need when running 'avbtool add_hashtree_footer' - do this
by adding an --calc_max_image_size option to the add_hashtree_footer
subcommand.

Previously we just added the hashtree and metadata at the beginning of
the DONT_CARE section, taking care to make a distinction between the
image size and the size we actually cared about. With this setup, this
extra accounting is no longer needed.

Bug: 31264226
Change-Id: Ide9dc8ef983b59d7a3dc9a1de179eaed8358dbd7
Test: New unit test and all unit tests pass.