Work around unpredictable output from make_ext4fs.
Since we're using the same external/avb tree on various branches we end
up using different versions of make_ext4fs which yield different binary
outputs given the same options. This is problematic as we currently rely
on a hash digest to have a given value. Because of this, the test
AvbToolTest.AddHashFooterSparseWithHoleAtTheEnd is passing on AOSP and
failing on goog/master.

Work around this by simply removing digest from what we're comparing
against.

Bug: None
Test: Unit tests pass on both aosp and goog/master.
Change-Id: I9c97e2bdb0ca9072535be2e59567fce8393221d6