Add room for expansion in descriptors.
As seen in the previous CL adding FEC data to the Hashtree descriptor,
it's not inconceivable that we'll need additional fields in the
future. Therefore add a reserved field of 64 bytes for each descriptor
type. This way we can add fields in the future without breaking forwards
or backwards compability.

Don't do this for Property or KernelCmdline descriptors since these are
already free-form data.

Bug: None
Test: Unit tests pass.
Change-Id: I4a277546fd48e83f861ab5484f7e18524b7b0117