Set androidboot.vbmeta.device from libavb instead of descriptor.
When using an initrd --generate_dm_verity_cmdline_from_hashtree is not
set because the initrd is expected to set up the root filesystem. In
this case androidboot.vbmeta.device still needs to be set but right
now this is not how things work. Fix this by setting this kernel
command-line parameter from within libavb instead of in a descriptor.

Bug: 33590159
Test: Unit tests pass.
Test: Manually tested on UEFI based bootloader.

Change-Id: I093c1b1f670082ca998f0a97a215676371ac0be2