avbtool: Make it possible to save vbmeta structs to a separate file.
This can be used when calculating hash and hashtree descriptors. For
some workloads this is desirable since it allows you to transfer the
much smaller vbmeta.bin file instead of the entire image. Also make it
possible to request that the vbmeta struct is not appended to the
image at all.

Bug: 34968890
Test: New unit tests and all unit tests pass.
Change-Id: I0dbdbd347d8c759460284bb0088dfd44fd5c90d0