libavb: Include androidboot.vbmeta.version in the generated command-line.
This version number represents the version of the libavb embedded in
the bootloader. This can be used at run-time to reject updates that
would write a vbmeta partition with a newer version than what the
bootloader supports.

(Also include some fixes for the previous CL which didn't make the cut
before Treehugger merged it.)

Bug: 35416772
Test: New unit tests and all unit tests pass.
Test: Manually tested on UEFI based bootloader.
Change-Id: Ie6aea68d0d5154be64cb1fff0699d16a0b123d39