avbtool: Add 'append_vbmeta_image' command.
This command is useful for when the target device has no vbmeta
partition and the vbmeta image is read from the boot partition instead.

When invoked, the vbmeta blob is copied bit-for-bit from the source
image and appended to the target image. After this, a footer is
appended to the target image.

Bug: 35880930
Test: New unit tests and all unit tests pass.
Test: Manually tested on UEFI based bootloader.
Change-Id: If524fff792c7c7420b11addb798260ee56fb0d98