avbtool: add_hashtree_footer: Add --setup_as_rootfs_from_kernel option.
This option makes it possible to generate system.img with kernel
command-line descriptors for setting up the partition with dm-verity
as the root filesystem. This is different from the existing option
--setup_rootfs_from_kernel which takes a path to system.img and adds
the kernel command-line descriptors to e.g. vbmeta.img.

This option can be used when using chained partitions for system.img.

Also fix README.md to use --include_descriptors_from_image and not
--include_descriptors_from_footer.

Bug: 38304536
Test: New unit tests + all unit tests pass.
Change-Id: I6285877cdb3b63a7c9117c270459d1fbb93e3309